add mapping of mob prefabs to config item names
Several mobs exist in variants (f. ex. treeguards, shadow chesspieces,
koalafant), but the mod configuration does not differentiate currently
to keep the config a bit leaner. When getting the configuration settings
for a specific mob, a map has been introduced to look up the config
item name for prefabs, where the prefab isn't identical to the config
item name.

Fixes #2.

Signed-off-by: Dennis Herbrich <kontakt+github@veloxis.de>