add log_level configuration option and Set/GetLogLevel()
The default log level of INFO can now be changed by a newly introduced
configuration option "log_level" that accepts the string log level
values TRACE, INFO, DEBUG, WARNING, and ERROR. The log level is
set automatically when requiring the logging module.

If the log_level configuration option does not exist, or contains an
invalid or unknown log level value, the log level will be unchanged, or
set to the default MOD_LOGLEVEL constant if not set before, making
exposure of the configuration option entirely optional.

Signed-off-by: Dennis Herbrich <kontakt+github@veloxis.de>