refactor logging module, clean const module
The lib.logging module is modularized and its dependency on the const
module removed. The modname is given to the Logging class ctor as an
argument instead of globalizing the modname variable across modules.

The lib.const module was subsequently cleaned of any logging-specific
settings, which have been moved into the logging module, as well as
unused or strictly unnecessary "helper" constants based on module
information, which should be injected as needed, when needed, instead.

Signed-off-by: Dennis Herbrich <kontakt+github@veloxis.de>