revert logging module to dependency on modname global
The logging module now expects _G.modname to be set in the environment,
not on it being set as a ctor argument. Certain "globals" like modname
and modinfo only exist in the modmain.lua script environment, and
`require`d modules generally do not have access to these variables, and
subsequently could not instantiate a Logging object.

Now, the modname required by the logging module is globalized for all
modules to use, if they need it, and the logging module itself
immediately returns a fully usable Logging instance, which seems more
intuitive overall.

Signed-off-by: Dennis Herbrich <kontakt+github@veloxis.de>