exempt example Discord webhook from validity check
Do not consider the webhook URL used in the documentation as a configuration
example as syntactically valid, and subsequently do not attempt to validate the
known invalid URL with Discord.  This prevents bogus traffic to the Discord
servers from people who are just following the instructions and not changing
the webhook URL in the example config.

Related to GH-15.